・KeepAliveの時間制御をすると間隔が短くてもKeepAliveに失敗するので頻度制御は全て削除 #78

+++ headtracker/rift.h
@@ -17,7 +17,7 @@ private:
 	const int fd;
 	static const int VendorID = 0x2833;
 	static const int ProductID = 0x0001;
-	static const int keepAliveInterval = 1000;
+	static const long keepAliveInterval = 1000;
 	static int OpenDevice();
 
 	// SENSOR
